requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2
Jinja2==2.10.3
Flask-Bcrypt==0.7.1
python-dotenv==0.10.3
plotly==4.3.0
pandas==0.25.3
dash==1.6.0
dash-daq==0.3.0
sshtunnel==0.1.5
SQLAlchemy==1.3.11
psycopg2-binary==2.8.4
python-dotenv==0.10.3
feedparser==5.2.1
beautifulsoup4==4.8.1